+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Implement image encoding for GIF.
 * Optimize image encoding, now is faster and uses less memory.
 * Add `IpcBytesMut::from_bytes` and `from_bytes_blocking`.
 * Add `IpcBytesMut::from_slice_blocking`.